### 2023-01-06 #5044 @editorialbot
[PRE REVIEW]: CeRULEo: Comprehensive utilitiEs for Remaining Useful Life Estimation methOds
**Submitting author:** @lucianolorenti (<a href="http://orcid.org/0000-0002-4041-449X">Luciano Rolando Lorenti</a>)
**Repository:** https://github.com/lucianolorenti/ceruleo
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.3
