### 2023-01-06 5aadd16
JOSS Paper: Add doi to items in the bibliography
