+++ paper/paper.bib
@@ -6,8 +6,10 @@
   number={3},
   pages={812--820},
   year={2014},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tii.2014.2349359}
 }
+
 
 @inproceedings{heimes2008recurrent,
   title={Recurrent neural networks for remaining useful life estimation},
@@ -15,7 +17,8 @@
   booktitle={2008 international conference on prognostics and health management},
   pages={1--6},
   year={2008},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/phm.2008.4711422}
 }
 
 
@@ -27,7 +30,8 @@
   volume={172},
   pages={1--11},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ress.2017.11.021}
 }
 
   @incollection{damant2021exploring,
@@ -36,7 +40,8 @@
   booktitle={Transdisciplinary Engineering for Resilience: Responding to System Disruptions},
   pages={171--180},
   year={2021},
-  publisher={IOS Press}
+  publisher={IOS Press},
+  doi={10.3233/atde210095}
 }
 
 
@@ -47,7 +52,8 @@
   journal={Applied Intelligence},
   pages={1--31},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10489-021-03004-y}
 }
 
 @inproceedings{rastegari2015cost,
@@ -56,7 +62,8 @@
   booktitle={2015 Annual Reliability and Maintainability Symposium (RAMS)},
   pages={1--6},
   year={2015},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/rams.2015.7105079}
 }
 
 @inproceedings{lemhadri2021lassonet,
@@ -65,7 +72,8 @@
   booktitle={International Conference on Artificial Intelligence and Statistics},
   pages={10--18},
   year={2021},
-  organization={PMLR}
+  organization={PMLR},
+  doi={10.48550/arXiv.1907.12207}
 }
 
 @article{fauvel2021xcm,
@@ -76,7 +84,8 @@
   number={23},
   pages={3137},
   year={2021},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/math9233137}
 }
 
 @inproceedings{jayasinghe2019temporal,
@@ -85,7 +94,8 @@
   booktitle={2019 IEEE International Conference on Industrial Technology (ICIT)},
   pages={915--920},
   year={2019},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icit.2019.8754956}
 }
 
 @article{li2020remaining,
@@ -95,7 +105,8 @@
   volume={89},
   pages={106113},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.asoc.2020.106113}
 }
 
 @article{CHEN2022104969,
