### 2023-01-06 3997480
JOSS Paper: Fix invalid DOI
