+++ paper/paper.bib
@@ -116,7 +116,7 @@ volume = {118},
 pages = {104969},
 year = {2022},
 issn = {0967-0661},
-doi = {https://doi.org/10.1016/j.conengprac.2021.104969},
+doi = {10.1016/j.conengprac.2021.104969},
 url = {https://www.sciencedirect.com/science/article/pii/S096706612100246X},
 author = {Yongyi Chen and Dan Zhang and Wen-an Zhang},
 keywords = {Rolling bearing, Remaining useful life estimation, Deep learning, Long-term recurrent convolutional network, Attentional mechanism},
