+++ setup.cfg
@@ -7,7 +7,7 @@ keywords = predictive maintenance, remaining useful life
 [options]
 packages =
     ceruleo
-python_requires = >=3.
+python_requires= >=3.7
 install_requires = 
     pyts
     pandas
