### 2023-03-15 4e95296
setup.cfg: Fix python_requires
