### 2023-04-11 lucianolorenti/ceruleo#23 @ulf1
Installation failed and dependencies have no versions 
## no versions
https://github.com/lucianolorenti/ceruleo/blob/4e952966dbc3389e0e1f1cd7635c81eeccfbec12/setup.cfg#L12

## current installation error
Command
```
pip install ceruleo
```

Error
```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
tensorflow 2.12.0 requires numpy<1.24,>=1.22, but you have numpy 1.21.6 which is incompatible.
```
