### 2023-04-12 0a2116c
Setup: Move everyhting to pyproject.toml and use hatchling for packaging
