+++ pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
-requires = ["setuptools", "setuptools-scm"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
 [project]
 name = "ceruleo"
@@ -9,34 +9,52 @@ readme = "README.md"
 requires-python = ">=3.7"
 keywords = ["predictive maintenance", "remaining useful life"]
 license = {text = "MIT"}
+dynamic = ["version"]
 classifiers = [
     "Programming Language :: Python :: 3",
     "Topic :: Scientific/Engineering :: Artificial Intelligence"
 ]
 dependencies = [
-        "pandas",
-        "numpy",
-        "tqdm",
-        "scikit-learn",
-        "gwpy",
-        "emd",
-        "dill",
-        "mmh3",
-        "pyarrow",
-        "fastparquet",
-        "sphinxcontrib.bibtex",
-        "gdown",
-        "uncertainties",
-        "pyinform",
-        "pyts",
-        "PyWavelets",
-        "seaborn",
-        "antropy"
-
+        "pandas >= 1.4",
+        "numpy >= 1.22",
+        "tqdm >= 4.56",
+        "scikit-learn >= 0.24",
+        "emd >= 0.4",
+        "mmh3   >= 2.0",
+        "pyarrow >= 4",
+        "gdown >= 4.2",
+        "pyinform >= 0.2",
+        "pyts >= 0.12",
+        "seaborn >= 0.11",
+        "antropy >= 0.1",
+        "uncertainties >= 3.1",
+        "PyWavelets >= 1.3",
 ]
-dynamic = ["version"]
+
+
+[tool.hatch.version]
+path = "ceruleo/__init__.py"
+
+
+[project.urls]
+Homepage = "https://github.com/lucianolorenti/ceruleo"
+Documentation = "https://lucianolorenti.github.io/ceruleo/"
 
 [project.optional-dependencies]
-doc = ["mkdocs"]
+tensorflow = ["tensorflow >= 2.5"]
 
-
+test = [
+    "pytest",
+    "coverage",
+    "xgboost >= 1.5"
+]
+doc = [
+    "mkdocs", 
+    "mkdocstrings[python]",
+    "mkdocs-material",
+    "mkdocs-jupyter",
+    "jupyter_contrib_nbextensions",
+    "mkdocstrings",
+    "mkdocs-bibtex",
+    "sphinxcontrib.bibtex"
+]
