### 2023-04-12 7d098d2
Setup: Use 1.5 as lowerbound for pandas
