+++ pyproject.toml
@@ -15,7 +15,7 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Artificial Intelligence"
 ]
 dependencies = [
-        "pandas >= 1.4",
+        "pandas >= 1.5",
         "numpy >= 1.22",
         "tqdm >= 4.56",
         "scikit-learn >= 0.24",
