### 2023-04-12 5f77c12
Tests: Update tests to avoid using skipna
