### 2023-04-12 85d8f8b
Improve setup (#24)
* Setup: Move everything to pyproject.toml and use hatchling for packaging

* Bump version

* Setup: Use 1.5 as lower bound for pandas

* Tests: Update tests to avoid using skipna
