### 2023-04-12 lucianolorenti/ceruleo#23 @lucianolorenti
Hi! I uploaded a new version to pypi with an improved setup and pinned libraries. Let me know if works now
