### 2023-04-18 #5294 @matthewfeickert
@ulf1 You've added helpful comments and notes to your review https://github.com/openjournals/joss-reviews/issues/5294#issuecomment-1484532861, which is exactly what we hope that reviewers will do. :+1: Some of your notes though point out problems or typos in the submission. Can you please translate all of those problems to GitHub Issues on https://github.com/lucianolorenti/cerule like you did with https://github.com/lucianolorenti/ceruleo/issues/23 (maybe prefixing the titles with [JOSS Review] to help @lucianolorenti differentiate them) ?
