### 2023-05-02 lucianolorenti/ceruleo#25 @AnnikaStein
[JOSS Review] Comments on the software paper (writing, references)
- Similar point as what has been raised by ulf, l.40: studyng -> studying
- The abbreviation PdM has been introduced, but has not been used once (compared to RUL, which is used several times throughout the paper). Consider using PdM or not introduce the abbreviation PdM.
- Syntax is fine, there could be additional citation of the software (l.44: scikit-learn, tensorflow)

https://github.com/openjournals/joss-reviews/issues/5294
