### 2023-05-07 b866287
JOSS: Update paper based on #25
Fixed typo on l.40: studyng -> studying
Use The abbreviation PdM
Add c iting for scikit-learn, tensorflow
