+++ paper/paper.bib
@@ -122,3 +122,63 @@ author = {Yongyi Chen and Dan Zhang and Wen-an Zhang},
 keywords = {Rolling bearing, Remaining useful life estimation, Deep learning, Long-term recurrent convolutional network, Attentional mechanism},
 abstract = {Rolling bearings are important components of industrial rotating machinery and equipment. The prediction of the remaining useful life (RUL) of rolling bearings is of great significance for improving the safety of the machine, reducing the economic and property losses caused by the failure of the bearings. However, for the task of predicting the RUL of rolling bearings, the information of the past time and the future time are as important as the information of the current time. In order to make better use of the extracted features for RUL prediction of rolling bearings, this paper has proposed a novel deep learning framework of multi-scale long-term recurrent convolutional network with wide first layer kernels and residual shrinkage building unit (MSWR-LRCN). The major difference from the previous deep neural network is that our new network organically combines the attention mechanism with multi-scale feature fusion strategy, and improves the anti-noise ability of the entire network. In addition, moving average (MA) method and a polynomial fitting model are also used, which help predict the RUL of rolling bearings effectively. The results show that this method has improved the prediction accuracy compared with the existing methods.}
 }
+
+@article{scikit-learn,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
+}
+
+@misc{tensorflow2015-whitepaper,
+title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+url={https://www.tensorflow.org/},
+note={Software available from tensorflow.org},
+author={
+    Mart\'{i}n~Abadi and
+    Ashish~Agarwal and
+    Paul~Barham and
+    Eugene~Brevdo and
+    Zhifeng~Chen and
+    Craig~Citro and
+    Greg~S.~Corrado and
+    Andy~Davis and
+    Jeffrey~Dean and
+    Matthieu~Devin and
+    Sanjay~Ghemawat and
+    Ian~Goodfellow and
+    Andrew~Harp and
+    Geoffrey~Irving and
+    Michael~Isard and
+    Yangqing Jia and
+    Rafal~Jozefowicz and
+    Lukasz~Kaiser and
+    Manjunath~Kudlur and
+    Josh~Levenberg and
+    Dandelion~Man\'{e} and
+    Rajat~Monga and
+    Sherry~Moore and
+    Derek~Murray and
+    Chris~Olah and
+    Mike~Schuster and
+    Jonathon~Shlens and
+    Benoit~Steiner and
+    Ilya~Sutskever and
+    Kunal~Talwar and
+    Paul~Tucker and
+    Vincent~Vanhoucke and
+    Vijay~Vasudevan and
+    Fernanda~Vi\'{e}gas and
+    Oriol~Vinyals and
+    Pete~Warden and
+    Martin~Wattenberg and
+    Martin~Wicke and
+    Yuan~Yu and
+    Xiaoqiang~Zheng},
+  year={2015},
+}
