+++ README.md
@@ -34,6 +34,18 @@ cd ceruleo
 pip install .
 ```
 
+# Contributing
+
+We love your input! We want to make contributing to this project as easy and transparent as possible, whether it's:
+
+* Reporting a bug
+* Discussing the current state of the code
+* Submitting a fix
+* Proposing new features
+
+Please feel free to open an [issue](https://github.com/fzi-forschungszentrum-informatik/TSInterpret/issues/new/choose) if you have any questions, find any bugs, or have any ideas.  
+Further information on how to contribute can be found by clicking  [here](https://github.com/lucianolorenti/ceruleo/blob/main/CONTRIBUTING.md).
+
 
 
 
