### 2023-05-14 02600e8
README: Add link to the contributing document in the readme (#27)
* README: Add link to the contributing document in the readme
