### 2023-05-19 #5294 @matthewfeickert
Hi All, just following up on the state of the review.

* @AnnikaStein, your checklist is complete and your https://github.com/lucianolorenti/ceruleo/issues/25 was resolved, so is your review complete and if so, what is your review recommendation?
* @Athene-ai, I believe your are done and your review recommendation (given no comments of any type) is to recommend publication?
* @ulf1 Can you update us on the state of things given your review checklist, the Issues you've opened (https://github.com/lucianolorenti/ceruleo/issues/23 has been resolved), and https://github.com/openjournals/joss-reviews/issues/5294#issuecomment-1546982158?
