### 2023-05-19 #5294 @Athene-ai
|I believe your are done and your review recommendation (given no comments
of any type) is to recommend publication?|

Yes, I recommend for the publication

On Fri, 19 May 2023 at 20:01, Matthew Feickert ***@***.***>
wrote:

> Hi All, just following up on the state of the review.
>
>    - @AnnikaStein <https://github.com/AnnikaStein>, your checklist is
>    complete and your lucianolorenti/ceruleo#25
>    <https://github.com/lucianolorenti/ceruleo/issues/25> was resolved, so
>    is your review complete and if so, what is your review recommendation?
>    - @Athene-ai <https://github.com/Athene-ai>, I believe your are done
>    and your review recommendation (given no comments of any type) is to
>    recommend publication?
>    - @ulf1 <https://github.com/ulf1> Can you update us on the state of
>    things given your review checklist, the Issues you've opened (
>    lucianolorenti/ceruleo#23
>    <https://github.com/lucianolorenti/ceruleo/issues/23> has been
>    resolved), and #5294 (comment)
>    <https://github.com/openjournals/joss-reviews/issues/5294#issuecomment-1546982158>
>    ?
>
