### 2023-05-19 #5294 @AnnikaStein
> Hi All, just following up on the state of the review.
>
> * @AnnikaStein, your checklist is complete and your [[JOSS Review] Comments on the software paper (writing, references) lucianolorenti/ceruleo#25](https://github.com/lucianolorenti/ceruleo/issues/25) was resolved, so is your review complete and if so, what is your review recommendation?
> * @Athene-ai, I believe your are done and your review recommendation (given no comments of any type) is to recommend publication?
> * @ulf1 Can you update us on the state of things given your review checklist, the Issues you've opened ([Installation failed and dependencies have no versions  lucianolorenti/ceruleo#23](https://github.com/lucianolorenti/ceruleo/issues/23) has been resolved), and [[REVIEW]: CeRULEo: Comprehensive utilitiEs for Remaining Useful Life Estimation methOds #5294 (comment)](https://github.com/openjournals/joss-reviews/issues/5294#issuecomment-1546982158)?

Hi Matthew,

indeed, my review is complete and I recommend this submission for publication.
