### 2023-05-23 #5294 @ulf1
>
> * Regarding the State of the field:
>
> While there are numerous repositories with published and unpublished models for remaining useful life regression, and other PdM tasks,  as well as multiple libraries for feature extraction in time series data, it is hard to find ibraries that combine both aspects, focusing on predictive maintenance and also including tools for model comparison in this context.
>
> To the best of my knolwedge, the closest library in spirit is the prog_models and prog_als libraries from NASA (https://github.com/nasa/prog_models), which are intended to be used together. These libraries have a broader scope, as they allow simulation, but they don't provide so many mechanisms for feature extraction from time series data. They include LSTM models and Dynamic Mode Decomposition wrapped in their own library-specific classes, to be used, i think, with the third component: the prog_server library. In the case of CeRULEo, it focuses on a data-driven approach using industrial data, where a clear simulation model may not be available or is expensive to develop. One of the guiding principles of the library is to avoid adding additional classes wrapping the models themselves, making them easily deployable in any production environment.  The idea is to  as model library-agnostic as possible.
>
>
> Do you think should I add some more paragraph to the paper.md regarding the libraries?

@matthewfeickert Should the NASA software quoted in the paper?
