### 2023-05-24 #5294 @matthewfeickert
> Respect to the CLA, I prefer not to use one.

There is no need to ask for a CLA.

> Do you think should I add some more paragraph to the paper.md regarding the libraries?
> ...
> Should the NASA software quoted in the paper?

I will leave it up to the authors what they feel as necessary to include in the paper as long as they are not excluding previous work in the area. However, given that this is a question that has come up I would suggest that some (short!) further clarifying discussion be added to clarify the position/scope of the library in the ecosystem.

@lucianolorenti as I just read a chunk of the paper to answer the above question I'll give you a heads up that in its current state of https://github.com/lucianolorenti/ceruleo/blob/02600e88a24f1b5ed982fe362f7664285b525870/paper/paper.md?plain=1#L32 "Industry 4.0" is jargon that will need to be updated.
