+++ paper/paper.md
@@ -43,6 +43,10 @@ In the context of predictive maintenance, explainability is crucial. As such, `C
 Moreover, `CeRULEo` provides tools for evaluating and comparing PdM models based on not only traditional regression metrics, but also on their ability to prevent errors and reduce costs. In many cases, the costs of not accurately detecting or anticipating faults can be much higher than the cost of inspections or maintenance due to reduced efficiency, unplanned downtime, and corrective maintenance expenses. In predictive maintenance, it is particularly important to be accurate about the remaining useful life  of equipment near the end of its lifespan, as an overestimation of RUL can have serious consequences when immediate action is required. `CeRULEo` addresses this issue by providing mechanisms for weighting samples according to their importance and asymmetric losses for training models, as well as visualization tools for understanding model performance in relation to true RUL.
 
 
+In the context of remaining useful life regression and other predictive maintenance (PdM) tasks, there exists a wide range of repositories containing published and unpublished models, as well as multiple libraries dedicated to feature extraction in time series data. However, it is challenging to find libraries that effectively combine both aspects, focusing on predictive maintenance while also incorporating tools for model comparison.
+
+To the best of my knowledge, the prog_models and prog_als libraries developed by NASA [@2022_nasa_prog_models] come closest to fulfilling these requirements. Although these libraries have a broader scope, including simulation capabilities, they lack extensive mechanisms for feature extraction from time series data. They do provide LSTM models and Dynamic Mode Decomposition, wrapped in their own library-specific classes, which can be utilized in conjunction with the prog_server library. Conversely, CeRULEo primarily emphasizes a data-driven approach using industrial data, particularly in cases where a clear simulation model may be unavailable or expensive to develop. Notably, the CeRULEo library adheres to the guiding principle of avoiding additional classes that wrap the models themselves, enabling seamless deployment in any production environment. The objective is to ensure maximum model library-agnosticism.
+
 # Financial Acknowledgement
 
 The Italian Government PNRR iniatiatives 'Partenariato 11: Made in Italy circolare e sostenibile' and 'Ecosistema dell'Innovazione - iNest' are gratefully acknowledged for partially financing this research activity.
