+++ paper/paper.bib
@@ -182,3 +182,12 @@ author={
     Xiaoqiang~Zheng},
   year={2015},
 }
+
+@misc{2022_nasa_prog_models,
+    author    = {Christopher Teubert and Matteo Corbetta and Chetan Kulkarni and Katelyn Jarvis and Matthew Daigle},
+    title     = {Prognostics Models Python Package},
+    month     = December,
+    year      = 2022,
+    version   = {1.4},
+    url       = {https://github.com/nasa/prog\_models}
+    }
