### 2023-05-30 daa9f04
Paper: Shorten the text describing the sota in libraries
