+++ paper/paper.bib
@@ -191,3 +191,27 @@ author={
     version   = {1.4},
     url       = {https://github.com/nasa/prog\_models}
     }
+
+  @article{christ2018time,
+  title={Time series feature extraction on basis of scalable hypothesis tests (tsfresh--a python package)},
+  author={Christ, Maximilian and Braun, Nils and Neuffer, Julius and Kempa-Liehr, Andreas W},
+  journal={Neurocomputing},
+  volume={307},
+  pages={72--77},
+  year={2018},
+  publisher={Elsevier}
+}
+
+@article{JMLR:v21:20-091,
+  author  = {Romain Tavenard and Johann Faouzi and Gilles Vandewiele and
+             Felix Divo and Guillaume Androz and Chester Holtz and
+             Marie Payne and Roman Yurchak and Marc Ru{\ss}wurm and
+             Kushal Kolar and Eli Woods},
+  title   = {Tslearn, A Machine Learning Toolkit for Time Series Data},
+  journal = {Journal of Machine Learning Research},
+  year    = {2020},
+  volume  = {21},
+  number  = {118},
+  pages   = {1-6},
+  url     = {http://jmlr.org/papers/v21/20-091.html}
+}
