### 2023-05-30 #5294 @matthewfeickert
> Assuming that the changes are made in the paper, I recommend the software for the publication.

Great! @ulf1 @lucianolorenti can one of you please open up a corresponding Issue for the remaining edits on https://github.com/lucianolorenti/ceruleo/issues so that this is tracked through both the Issue tracker and this review?
