### 2023-06-01 lucianolorenti/ceruleo#29 @lucianolorenti
JOSS: Scope of the library in the ecosystem and Incipient Industry 5.0 
Two comments were made in the JOSS review process about clarifying the scope of the library in the current ecosystem, and about the mention of industry 4.0, which with the advent of its evolution, the industry 5.0, is a bit outdated

The PR covering this is #28
