### 2023-06-01 #5294 @lucianolorenti
Here is the issue  lucianolorenti/ceruleo#29

I started working on the changes to the paper, I will try to finish ASAP. This week has been a bit complicated
