+++ ceruleo/transformation/features/extraction.py
@@ -20,6 +20,8 @@ from ceruleo.transformation import TransformerStep
 from ceruleo.transformation.features.rolling_windows import apply_rolling_data
 from ceruleo.transformation.functional.transformers import Transformer
 from ceruleo.transformation.utils import SKLearnTransformerWrapper
+from pandas.core.window.rolling import Rolling
+
 
 logger = logging.getLogger(__name__)
 
@@ -244,10 +246,8 @@ class SimpleEncodingCategorical(TransformerStep):
         return pd.DataFrame({"encoding": d.codes}, index=X.index)
 
 
-
-
 def rolling_kurtosis(s: pd.Series, window, min_periods):
-    return s.rolling(window, min_periods=min_periods).kurt(skipna=True)
+    return s.rolling(window, min_periods=min_periods).kurt(numeric_only=True)
 
 
 class LifeStatistics(TransformerStep):
@@ -285,9 +285,8 @@ class LifeStatistics(TransformerStep):
     """
 
     def __init__(
-        self, *,to_compute: Optional[List[str]] = None, name: Optional[str] = None
+        self, *, to_compute: Optional[List[str]] = None, name: Optional[str] = None
     ):
-
         super().__init__(name=name)
         valid_stats = [
             "kurtosis",
@@ -459,7 +458,7 @@ class RollingStatistics(TransformerStep):
         ]
 
         if to_compute is not None and specific is not None:
-            raise ValueError('Only one of to_compute or specific should be used')
+            raise ValueError("Only one of to_compute or specific should be used")
         self.specific = specific
         self.to_compute = to_compute
         if to_compute is None:
@@ -483,67 +482,71 @@ class RollingStatistics(TransformerStep):
 
     def _std_asinh(self, X, rolling, abs_rolling):
         return (
-            X.apply(np.arcsinh).rolling(self.window, self.min_points).std(skipna=True)
+            X.apply(np.arcsinh).rolling(self.window, self.min_points).std(numeric_only=True)
         )
 
     def _std_acosh(self, X, rolling, abs_rolling):
         return (
-            X.apply(np.arccosh).rolling(self.window, self.min_points).std(skipna=True)
+            X.apply(np.arccosh).rolling(self.window, self.min_points).std(numeric_only=True)
         )
 
     def _energy(self, X, rolling, abs_rolling):
         return X.pow(2).rolling(self.window, self.min_points).sum()
 
     def _std_atan(self, X, rolling, abs_rolling):
-        return X.apply(np.arctan).rolling(self.window, self.min_points).std(skipna=True)
+        return (
+            X.apply(np.arctan)
+            .rolling(self.window, self.min_points)
+            .std(numeric_only=True)
+        )
 
-    def _mean(self, X, rolling, abs_rolling):
-        return rolling.mean(skipna=True)
+    def _mean(self, X, rolling: Rolling, abs_rolling: Rolling):
+        return rolling.mean(numeric_only=True)
 
-    def _kurtosis(self, X, rolling, abs_rolling):
-        return rolling.kurt(skipna=True)
+    def _kurtosis(self, X, rolling: Rolling, abs_rolling: Rolling):
+        return rolling.kurt(numeric_only=True)
 
-    def _skewness(self, X, rolling, abs_rolling):
-        return rolling.skew(skipna=True)
+    def _skewness(self, X, rolling: Rolling, abs_rolling: Rolling):
+        return rolling.skew(numeric_only=True)
 
-    def _max(self, X, rolling, abs_rolling):
-        return rolling.max(skipna=True)
[…]
