+++ setup.cfg
@@ -1,42 +1,0 @@
-[metadata]
-name = ceruleo
-version = 2.0.3
-description = Remaining useful life estimation utilities
-url = https://github.com/lucianolorenti/ceruleo
-keywords = predictive maintenance, remaining useful life
-[options]
-packages =
-    ceruleo
-python_requires= >=3.7
-install_requires = 
-    pyts
-    pandas
-    numpy
-    tqdm
-    scikit-learn
-    gwpy
-    emd
-    dill
-    mmh3
-    pyarrow
-    fastparquet
-    gdown
-    uncertainties
-    pyinform
-    pyts
-    PyWavelets
-    seaborn
-    antropy
-
-[options.extras_require]
-test = 
-    pytest
-    coverage
-doc = 
-    mkdocs 
-    mkdocstrings[python]
-    mkdocs-material
-    mkdocs-jupyter
-    jupyter_contrib_nbextensions
-    mkdocstrings
-    mkdocs-bibtex    
