### 2023-06-09 c5c63c6
Paper: Comment on Industry 5.0 and add appropiate references
