+++ paper/paper.bib
@@ -215,3 +215,23 @@ author={
   pages   = {1-6},
   url     = {http://jmlr.org/papers/v21/20-091.html}
 }
+
+@article{van2022predictive,
+  title={Predictive maintenance for industry 5.0: behavioural inquiries from a work system perspective},
+  author={van Oudenhoven, Bas and Van de Calseyde, Philippe and Basten, Rob and Demerouti, Evangelia},
+  journal={International Journal of Production Research},
+  pages={1--20},
+  year={2022},
+  publisher={Taylor \& Francis}
+}
+
+@article{khan2023changes,
+  title={Changes and improvements in Industry 5.0: A strategic approach to overcome the challenges of Industry 4.0},
+  author={Khan, Moin and Haleem, Abid and Javaid, Mohd},
+  journal={Green Technologies and Sustainability},
+  volume={1},
+  number={2},
+  pages={100020},
+  year={2023},
+  publisher={Elsevier}
+}
