### 2023-06-10 #5294 @lucianolorenti
Hi!
I updated the article in the [sota_libraries](https://github.com/lucianolorenti/ceruleo/tree/sota_libraries) branch addressing the issues raised.

You can read the PDF in the following action [result](https://github.com/lucianolorenti/ceruleo/actions/runs/5231736715)

Please, let me know what you think
