### 2023-06-10 99c5190
Paper: Remove repeated word
