### 2023-07-04 #5294 @matthewfeickert
@lucianolorenti Apologies for the delay, I've been travelling for work for the last several weeks, but I think this submission is very close to publication. Can you please revise the jargon I mentioned in https://github.com/openjournals/joss-reviews/issues/5294#issuecomment-1560521972 ?

Beyond that, all that's left to do now before publication is to ensure that there is a long term public archive of the code that was reviewed. In this case the code was [`v2.3.0`](https://github.com/lucianolorenti/ceruleo/releases/tag/v2.0.3) though there has been [some development between then and the current `HEAD`](https://github.com/lucianolorenti/ceruleo/compare/v2.0.3...02600e88a24f1b5ed982fe362f7664285b525870).

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

If you'd like to make a new release of CeRULEo to trigger the uploader that's fine and we can have @editorialbot update the version listed in the review. You're also welcome to not do that and just upload the state of the repository at [`v2.0.4`](https://github.com/lucianolorenti/ceruleo/releases/tag/v2.0.4). Either way we'll just need you to share the resulting DOI of the archive created in this thread.

Let me know if you have any questions. :+1:

@AnnikaStein @ulf1 Thank you very much for your reviews! I think they have both helped this submission reach a more improved state before publication.
