### 2023-07-05 #5294 @lucianolorenti
Hi @matthewfeickert , don't worry.
I made some changes regarding the jargon in https://github.com/openjournals/joss-reviews/issues/5294#issuecomment-1585797582

Let me know if that is what you meant
