### 2023-07-20 #5294 @matthewfeickert
@lucianolorenti Apologies for the delay &mdash; I've been traveling for work for the last 3 weeks and have been slow on returning to things.

> I made some changes regarding the jargon in https://github.com/openjournals/joss-reviews/issues/5294#issuecomment-1585797582

Thanks.

> In Industry 5.0, the industrial machines produce a large amount of data which can be used to predict an asset’s life [@khan2023changes]. RUL estimation uses prediction techniques to forecast a machine's future performance based on historical data, enabling early identification of potential failures and prompt pre-failure interventions.

It still seems a little jargony, but I'll accept it given the citation of Khan. :+1:  Though along the lines of citation I'm noticing that many of the references you give are missing the DOIs which are given on their journal page. A short example:

```bib
  @article{christ2018time,
  title={Time series feature extraction on basis of scalable hypothesis tests (tsfresh--a python package)},
  author={Christ, Maximilian and Braun, Nils and Neuffer, Julius and Kempa-Liehr, Andreas W},
  journal={Neurocomputing},
  volume={307},
  pages={72--77},
  year={2018},
  publisher={Elsevier}
}

...

@article{van2022predictive,
  title={Predictive maintenance for industry 5.0: behavioural inquiries from a work system perspective},
  author={van Oudenhoven, Bas and Van de Calseyde, Philippe and Basten, Rob and Demerouti, Evangelia},
  journal={International Journal of Production Research},
  pages={1--20},
  year={2022},
  publisher={Taylor \& Francis}
}

@article{khan2023changes,
  title={Changes and improvements in Industry 5.0: A strategic approach to overcome the challenges of Industry 4.0},
  author={Khan, Moin and Haleem, Abid and Javaid, Mohd},
  journal={Green Technologies and Sustainability},
  volume={1},
  number={2},
  pages={100020},
  year={2023},
  publisher={Elsevier}
}
```

all have DOIs posted. Can you please make sure that DOIs are included for all references that provide them?
