+++ paper/paper.bib
@@ -132,7 +132,8 @@ abstract = {Rolling bearings are important components of industrial rotating mac
  journal={Journal of Machine Learning Research},
  volume={12},
  pages={2825--2830},
- year={2011}
+ year={2011},
+ doi={https://doi.org/10.48550/arXiv.1201.0490}
 }
 
 @misc{tensorflow2015-whitepaper,
@@ -181,6 +182,7 @@ author={
     Yuan~Yu and
     Xiaoqiang~Zheng},
   year={2015},
+  doi={https://doi.org/10.48550/arXiv.1603.04467}
 }
 
 @misc{2022_nasa_prog_models,
@@ -199,7 +201,8 @@ author={
   volume={307},
   pages={72--77},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.neucom.2018.03.067}
 }
 
 @article{JMLR:v21:20-091,
@@ -222,7 +225,8 @@ author={
   journal={International Journal of Production Research},
   pages={1--20},
   year={2022},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={https://doi.org/10.1080/00207543.2022.2154403}
 }
 
 @article{khan2023changes,
@@ -233,5 +237,6 @@ author={
   number={2},
   pages={100020},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.grets.2023.100020} 
 }
