### 2023-07-23 0405532
Setup: Increase version of antropy due to error in CI
