### 2023-07-23 2b74ae2
Setup: Add upper bound for scikit-learn due to antropy incompatibility
