+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
         "pandas >= 1.5",
         "numpy >= 1.22",
         "tqdm >= 4.56",
-        "scikit-learn >= 0.24",
+        "scikit-learn >= 0.24, <1.3",
         "emd >= 0.4",
         "mmh3   >= 2.0",
         "pyarrow >= 4",
