### 2023-07-23 a3bd15c
Paper: Remove the URL from the new DOIs
