+++ paper/paper.bib
@@ -133,7 +133,7 @@ abstract = {Rolling bearings are important components of industrial rotating mac
  volume={12},
  pages={2825--2830},
  year={2011},
- doi={https://doi.org/10.48550/arXiv.1201.0490}
+ doi={10.48550/arXiv.1201.0490}
 }
 
 @misc{tensorflow2015-whitepaper,
@@ -182,7 +182,7 @@ author={
     Yuan~Yu and
     Xiaoqiang~Zheng},
   year={2015},
-  doi={https://doi.org/10.48550/arXiv.1603.04467}
+  doi={10.48550/arXiv.1603.04467}
 }
 
 @misc{2022_nasa_prog_models,
@@ -202,7 +202,7 @@ author={
   pages={72--77},
   year={2018},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.neucom.2018.03.067}
+  doi={10.1016/j.neucom.2018.03.067}
 }
 
 @article{JMLR:v21:20-091,
@@ -226,7 +226,7 @@ author={
   pages={1--20},
   year={2022},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.1080/00207543.2022.2154403}
+  doi={10.1080/00207543.2022.2154403}
 }
 
 @article{khan2023changes,
@@ -238,5 +238,5 @@ author={
   pages={100020},
   year={2023},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.grets.2023.100020} 
+  doi={10.1016/j.grets.2023.100020} 
 }
