### 2023-07-23 c12e24d
Add comment regarding the current state of libraries for PdM (#28)
* Paper: Add comment regarding the current state of libraries for PdM

* Paper: Comment on Industry 5.0 and add appropiate references

* Paper: Add missing DOIs

* Setup: Add upper bound for scikit-learn due to antropy incompatibility
