+++ paper/paper.bib
@@ -132,7 +132,8 @@ abstract = {Rolling bearings are important components of industrial rotating mac
  journal={Journal of Machine Learning Research},
  volume={12},
  pages={2825--2830},
- year={2011}
+ year={2011},
+ doi={10.48550/arXiv.1201.0490}
 }
 
 @misc{tensorflow2015-whitepaper,
@@ -181,4 +182,61 @@ author={
     Yuan~Yu and
     Xiaoqiang~Zheng},
   year={2015},
+  doi={10.48550/arXiv.1603.04467}
+}
+
+@misc{2022_nasa_prog_models,
+    author    = {Christopher Teubert and Matteo Corbetta and Chetan Kulkarni and Katelyn Jarvis and Matthew Daigle},
+    title     = {Prognostics Models Python Package},
+    month     = December,
+    year      = 2022,
+    version   = {1.4},
+    url       = {https://github.com/nasa/prog\_models}
+    }
+
+  @article{christ2018time,
+  title={Time series feature extraction on basis of scalable hypothesis tests (tsfresh--a python package)},
+  author={Christ, Maximilian and Braun, Nils and Neuffer, Julius and Kempa-Liehr, Andreas W},
+  journal={Neurocomputing},
+  volume={307},
+  pages={72--77},
+  year={2018},
+  publisher={Elsevier},
+  doi={10.1016/j.neucom.2018.03.067}
+}
+
+@article{JMLR:v21:20-091,
+  author  = {Romain Tavenard and Johann Faouzi and Gilles Vandewiele and
+             Felix Divo and Guillaume Androz and Chester Holtz and
+             Marie Payne and Roman Yurchak and Marc Ru{\ss}wurm and
+             Kushal Kolar and Eli Woods},
+  title   = {Tslearn, A Machine Learning Toolkit for Time Series Data},
+  journal = {Journal of Machine Learning Research},
+  year    = {2020},
+  volume  = {21},
+  number  = {118},
+  pages   = {1-6},
+  url     = {http://jmlr.org/papers/v21/20-091.html}
+}
+
+@article{van2022predictive,
+  title={Predictive maintenance for industry 5.0: behavioural inquiries from a work system perspective},
+  author={van Oudenhoven, Bas and Van de Calseyde, Philippe and Basten, Rob and Demerouti, Evangelia},
+  journal={International Journal of Production Research},
+  pages={1--20},
+  year={2022},
+  publisher={Taylor \& Francis},
+  doi={10.1080/00207543.2022.2154403}
+}
+
+@article{khan2023changes,
+  title={Changes and improvements in Industry 5.0: A strategic approach to overcome the challenges of Industry 4.0},
+  author={Khan, Moin and Haleem, Abid and Javaid, Mohd},
+  journal={Green Technologies and Sustainability},
+  volume={1},
+  number={2},
+  pages={100020},
+  year={2023},
+  publisher={Elsevier},
+  doi={10.1016/j.grets.2023.100020} 
 }
