### 2023-07-26 78974de
Setup: Put an upper bound on notebook to avoid issues with the documentation
