+++ .github/workflows/documentation.yml
@@ -26,7 +26,7 @@ jobs:
           python -m pip install --upgrade nbconvert
           python -m pip install  jinja2
           if [ -f requirements.txt ]; then python -m pip install -r requirements.txt; fi
-          python -m pip install -e $GITHUB_WORKSPACE
+          python -m pip install -e $GITHUB_WORKSPACE[doc]
 
 
       - name: Deploy docs
