### 2023-07-26 #5294 @lucianolorenti
Hi @matthewfeickert

I updated the missing DOIs and created a new release that is stored v2.0.5 that is stored in zenodo with the following DOI 10.5281/zenodo.8187300
