### 2023-08-03 #5294 @matthewfeickert
One final little fix @lucianolorenti! I just opened up https://github.com/lucianolorenti/ceruleo/pull/31 for this as the current paper url render is `https://github.com/nasa/prog/_models` which (correctly) 404s.
