### 2023-08-09 c54dcef
docs: Correct prog_models URL (#31)
* Correct URL to avoid trying to escape underscore so render in paper will be correct.
   - c.f. https://github.com/nasa/prog_models
