+++ paper/paper.bib
@@ -191,7 +191,7 @@ author={
     month     = December,
     year      = 2022,
     version   = {1.4},
-    url       = {https://github.com/nasa/prog\_models}
+    url       = {https://github.com/nasa/prog_models}
     }
 
   @article{christ2018time,
