### 2023-08-14 #5294 @gianantonio
Dear editors and reviewers,

thank you very much for handling the reviewing process of our work.

We realized that financial acknowledgements should be modified and we would like to ask for a post-publication update to the paper.

If possible the ack should be as follow:
"This work was partially carried out within the MICS (Made in Italy – Circular and Sustainable) Extended Partnership and received funding from Next-GenerationEU (Italian PNRR – M4 C2, Invest 1.3 – D.D. 1551.11-10-2022, PE00000004). Moreover this study was also partially carried out within the PNRR research activities of the consortium iNEST (Interconnected North-Est Innovation Ecosystem) funded by the European Union Next-GenerationEU (Piano Nazionale di Ripresa e Resilienza (PNRR) – Missione 4 Componente 2, Investimento 1.5 – D.D. 1058  23/06/2022, ECS00000043). This work was also co-funded by the European Union in the context of the Horizon Europe project 'AIMS5.0 - Artificial Intelligence in Manufacturing leading to Sustainability and Industry5.0' Grant agreement ID: 101112089."

Thank you for you support!
