### 2023-08-14 #5294 @kyleniemeyer
@gianantonio if you can update the paper source with that acknowledgment, I can reprocess the final PDF.
